pyproject.toml
[tool.poetry]
name = "chat-langchain"
version = "0.1.0"
description = ""
authors = ["CropGuard"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.28.0"
fastapi = "^0.103.1"
pydantic = "1.10"
langchain = "^0.0.333"
uvicorn = "^0.23.2"
beautifulsoup4 = "^4.12.2"
tiktoken = "^0.4.0"
weaviate-client = "^3.23.2"
psycopg2 = "^2.9.7"
lxml = "^4.9.3"
langserve = {extras = ["server"], version = "^0.0.34"}
voyageai = "^0.1.4"

python-dotenv = "1.0.0"
chromadb = "0.4.18"
sentence-transformers = "2.2.2"
lark = "1.1.8"
langsmith = "0.0.69"
langchain-core = "0.0.12"
# dsmain = "0.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

